move into entity_CRUD_blueprint
