+++ src/routes/entity_CRUD/__init__.py
@@ -8,10 +8,11 @@ import urllib.request
 import yaml
 from werkzeug import utils
 from operator import itemgetter
+from threading import Thread
 
 from hubmap_commons.hm_auth import AuthHelper
 from hubmap_commons.exceptions import HTTPException
-from hubmap_commons import file_helper as commons_file_helper
+from hubmap_commons import file_helper as commons_file_helper, neo4j_driver
 
 entity_CRUD_blueprint = Blueprint('entity_CRUD', __name__)
 logger = logging.getLogger(__name__)
@@ -256,6 +257,104 @@ def create_datasets_from_bulk():
             else:
                 entity_created = True
         return _send_response_on_file(entity_created, entity_failed_to_create, entity_response)
+
+
+@entity_CRUD_blueprint.route('/datasets/<uuid>/submit', methods=['PUT'])
+def submit_dataset(uuid):
+    if not request.is_json:
+        return Response("json request required", 400)
+    try:
+        dataset_request = request.json
+        auth_helper = AuthHelper.configured_instance(current_app.config['APP_CLIENT_ID'],
+                                                     current_app.config['APP_CLIENT_SECRET'])
+        ingest_helper = IngestFileHelper(current_app.config)
+        auth_tokens = auth_helper.getAuthorizationTokens(request.headers)
+        dataset_helper = DatasetHelper(current_app.config)
+        if isinstance(auth_tokens, Response):
+            return (auth_tokens)
+        elif isinstance(auth_tokens, str):
+            token = auth_tokens
+        # Do we need nexus token
+        elif 'nexus_token' in auth_tokens:
+            token = auth_tokens['nexus_token']
+        else:
+            return (Response("Valid nexus auth token required", 401))
+
+        if 'group_uuid' in dataset_request:
+            return Response(
+                "Cannot specify group_uuid.  The group ownership cannot be changed after an entity has been created.",
+                400)
+
+        group_uuid = dataset_helper.get_group_uuid_by_dataset_uuid(uuid)
+
+        user_info = auth_helper.getUserInfo(token, getGroups=True)
+        if isinstance(user_info, Response):
+            return user_info
+        if not 'hmgroupids' in user_info:
+            return Response("user not authorized to submit data, unable to retrieve any group information", 403)
+        if not current_app.config['SENNET_DATA_ADMIN_GROUP_UUID'] in user_info['hmgroupids']:
+            return Response("user not authorized to submit data, must be a member of the SenNet-Data-Admin group", 403)
+
+        # TODO: Temp fix till we can get this in the "Validation Pipeline"... add the validation code here... If it returns any errors fail out of this. Return 412 Precondition Failed with the errors in the description.
+        pipeline_url = commons_file_helper.ensureTrailingSlashURL(
+            current_app.config['INGEST_PIPELINE_URL']) + 'request_ingest'
+    except Exception as e:
+        logger.error(e, exc_info=True)
+        return Response("Unexpected error while creating a dataset: " + str(e) + "  Check the logs", 500)
+    try:
+        put_url = commons_file_helper.ensureTrailingSlashURL(
+            current_app.config['ENTITY_WEBSERVICE_URL']) + 'entities/' + uuid
+        dataset_request['status'] = 'Processing'
+        response = requests.put(put_url, json=dataset_request,
+                                headers={'Authorization': 'Bearer ' + token, 'X-SenNet-Application': 'ingest-api'},
+                                verify=False)
+        if not response.status_code == 200:
+            error_msg = f"call to {put_url} failed with code:{response.status_code} message:" + response.text
+            logger.error(error_msg)
+            return Response(error_msg, response.status_code)
+    except HTTPException as hte:
+        logger.error(hte)
+        return Response("Unexpected error while updating dataset: " + str(e) + "  Check the logs", 500)
+
+    def call_airflow():
+        try:
+            r = requests.post(pipeline_url, json={"submission_id": "{uuid}".format(uuid=uuid),
+                                                  "process": "SCAN.AND.BEGIN.PROCESSING",
+                                                  "full_path": ingest_helper.get_dataset_directory_absolute_path(
+                                                      dataset_request, group_uuid, uuid),
+                                                  "provider": "{group_name}".format(
+                                                      group_name=AuthHelper.getGroupDisplayName(group_uuid))},
+                              headers={'Content-Type': 'application/json', 'Authorization': 'Bearer {token}'.format(
+                                  token=AuthHelper.instance().getProcessSecret())}, verify=False)
+            if r.ok == True:
+                """expect data like this:
+                {"ingest_id": "abc123", "run_id": "run_657-xyz", "overall_file_count": "99", "top_folder_contents": "["IMS", "processed_microscopy","raw_microscopy","VAN0001-RK-1-spatial_meta.txt"]"}
+                """
+                data = json.loads(r.content.decode())
+                submission_data = data['response']
+                dataset_request['ingest_id'] = submission_data['ingest_id']
+                dataset_request['run_id'] = submission_data['run_id']
+            else:
+                error_message = 'Failed call to AirFlow HTTP Response: ' + str(r.status_code) + ' msg: ' + str(r.text)
+                logger.error(error_message)
+                dataset_request['status'] = 'Error'
+                dataset_request['pipeline_message'] = error_message
+            response = requests.put(put_url, json=dataset_request,
+                                    headers={'Authorization': 'Bearer ' + token, 'X-SenNet-Application': 'ingest-api'},
+                                    verify=False)
+            if not response.status_code == 200:
+                error_msg = f"call to {put_url} failed with code:{response.status_code} message:" + response.text
+                logger.error(error_msg)
+            else:
+                logger.info(response.json())
+        except HTTPException as hte:
+            logger.error(hte)
+        except Exception as e:
+            logger.error(e, exc_info=True)
+
+    thread = Thread(target=call_airflow)
+    thread.start()
+    return Response("Request of Dataset Submisssion Accepted", 202)
 
 
 def _bulk_upload_and_validate(entity):
